requirements.txt
python-telegram-bot==20.3
youtube-transcript-api==0.6.2
python-dotenv==1.0.1